+++ MetadataServer.py
@@ -2,21 +2,9 @@
 # -*- coding: utf-8 -*-
 # pylint: disable=invalid-name
 """Main metadata server script."""
-from json import dumps
 import cherrypy
-from metadata import CHERRYPY_CONFIG, main
+from metadata import CHERRYPY_CONFIG, main, error_page_default
 from metadata.rest.root import Root
-
-
-def error_page_default(**kwargs):
-    """The default error page should always enforce json."""
-    cherrypy.response.headers['Content-Type'] = 'application/json'
-    return dumps({
-        'status': kwargs['status'],
-        'message': kwargs['message'],
-        'traceback': kwargs['traceback'],
-        'version': kwargs['version']
-    })
 
 
 cherrypy.config.update({'error_page.default': error_page_default})
